requirements.txt
python-dotenv==1.0.1
langchain==0.3.4
langchain-community==0.3.2
langchain-openai==0.2.3
chromadb==0.5.5
faiss-cpu==1.8.0.post1
tiktoken==0.7.0

duckduckgo-search==6.1.6
requests==2.32.3

streamlit==1.39.0
pypdf==4.3.1
docx2txt==0.8